// src/generated/growi.ts
export type ObjectId = string;

/**
 * Page path
 */
export type PagePath = string;

/**
 * Grant for page
 */
export type PageGrant = number;
// ...
export type RevisionBody = string;
// ...
export type PostPageBodyGrantUserGroupIdsItem = {
  /** Group type */
  type?: string;
  /** UserGroup ID */
  item?: string;
};

export type PostPageBody = {
  /** Text of page */
  body: string;
  path: PagePath;
  grant?: PageGrant;
  grantUserGroupIds?: PostPageBodyGrantUserGroupIdsItem[];
  pageTags?: string[];
};

export type PostPage201 = {
  page?: Page;
  tags?: Tags[];
  revision?: Revision;
};

export type PutPageBodyUserRelatedGrantUserGroupIdsItem = {
  /** Group type */
  type?: string;
  /** UserGroup ID */
  item?: string;
};

export type PutPageBody = {
  body: RevisionBody;
  pageId: ObjectId;
  revisionId: ObjectId;
  grant?: PageGrant;
  userRelatedGrantUserGroupIds?: PutPageBodyUserRelatedGrantUserGroupIdsItem[];
  /** Determine whether the scopes of descendants should be overwritten */
  overwriteScopesOfDescendants?: boolean;
  /** Determine whether the page is enabled to be posted to Slack */
  isSlackEnabled?: boolean;
  /** Slack channel IDs */
  slackChannels?: string;
  /** Origin is "view" or "editor" */
  origin?: string;
  /** Determine whether the page is WIP */
  wip?: boolean;
};